+++ Sources/MapboxSearch/PublicAPI/HistoryRecord.swift
@@ -25,6 +25,9 @@ public struct HistoryRecord: IndexableRecord, SearchResult, Codable, Hashable {
     /// Record's name
     public private(set) var name: String
     
+    /// Index in response from server.
+    public let serverIndex: Int?
+    
     /**
         The feature name, as matched by the search algorithm.
         
@@ -92,6 +95,7 @@ public struct HistoryRecord: IndexableRecord, SearchResult, Codable, Hashable {
         id: String = UUID().uuidString,
         name: String,
         matchingName: String?,
+        serverIndex: Int?,
         coordinate: CLLocationCoordinate2D,
         timestamp: Date = Date(),
         historyType: HistoryRecord.HistoryType,
@@ -104,6 +108,7 @@ public struct HistoryRecord: IndexableRecord, SearchResult, Codable, Hashable {
         self.id = id
         self.name = name
         self.matchingName = matchingName
+        self.serverIndex = serverIndex
         self.coordinateCodable = .init(coordinate)
         self.timestamp = timestamp
         self.historyType = historyType
@@ -127,6 +132,7 @@ public struct HistoryRecord: IndexableRecord, SearchResult, Codable, Hashable {
         self.id = searchResult.id
         self.name = searchResult.name
         self.matchingName = searchResult.matchingName
+        self.serverIndex = searchResult.serverIndex
         self.coordinateCodable = .init(searchResult.coordinate)
         self.timestamp = timestamp
         self.historyType = historyType