+++ Sources/MapboxSearch/PublicAPI/Search Results/SearchCategorySuggestionImpl.swift
@@ -14,6 +14,8 @@ class SearchCategorySuggestionImpl: SearchCategorySuggestion, CoreResponseProvid
     
     var iconName: String?
     
+    var serverIndex: Int?
+    
     var suggestionType: SearchSuggestType
     
     var categories: [String]?
@@ -32,6 +34,7 @@ class SearchCategorySuggestionImpl: SearchCategorySuggestion, CoreResponseProvid
         self.name = coreResult.names[0]
         self.address = coreResult.addresses?.first.map(Address.init)
         self.iconName = nil // Categories should use it's special icon
+        self.serverIndex = coreResult.serverIndex?.intValue
         self.originalResponse = CoreSearchResultResponse(coreResult: coreResult, response: response)
         self.distance = coreResult.distanceToProximity
         self.batchResolveSupported = coreResult.action?.isMultiRetrievable ?? false