+++ Tests/MapboxSearchTests/CodablePersistentServiceTests.swift
@@ -35,6 +35,7 @@ class CodablePersistentServiceTests: XCTestCase {
                                     coordinate: coordinate,
                                     address: address,
                                     makiIcon: nil,
+                                    serverIndex: nil,
                                     categories: [],
                                     resultType: .address(subtypes: [.address]))
         XCTAssertTrue(service.saveData(record), "Unable to save record")
@@ -53,6 +54,7 @@ class CodablePersistentServiceTests: XCTestCase {
         let record = HistoryRecord(id: UUID().uuidString,
                                    name: "DaName",
                                    matchingName: nil,
+                                   serverIndex: nil,
                                    coordinate: coordinate,
                                    timestamp: Date(),
                                    historyType: .category,